NavMesh does not fail to find a path which ends to closed area.
Navigation uses a simple euclidean distance as a heuristic measure for
the choosing the best end face if the path cannot be completed. There are
a few special cases that can cause errors if they occur.

When trying to start from blocked area, no path is found, but the navpath
is still constructed from the start and end points.
  - This is 'avoided' by trying to find a face close to the start face,
    that is active.