Added tiny poly partition to replace the buggy poly2tri.
Now using tiny poly partition (TPP) to triangulate the navmesh with simple
ear clipping algorithm. The ear clipping can result in bad triangles (long and
thin), so a refinement step is done to the resulting triangulation.

Improved the NavMesh::FindNodePath to use edge midpoints instead of face
centers to calculate the cost.