Fixed navmesh generation in release mode.
In release mode the vertex cache of navmesh did not consider two vertices
to be same if they differed even a little. This caused all of the
vertices to be unique and failing the face neighbour search that relies on
the vertex cache merging two vertices to one if they are too close to each
other.