Fixed: tiny poly partition failing on certain cases.
As an attempt to fix tiny poly partition failing on certain input, when
trying to remove holes from the polygons, moved rightmost hole vertex to
left by 0.01. This should work as a workaround to the problem in the
context of Chocolate Cake as no holes that are less than 0.01 in width
should ever be generated (hause width + agent radius >> 0.01).

Some other changes were made.