Fix: find node path not including the start face.
Fix: navmesh generation failing due to Tiny Poly Partition not
understanding overlapping geometry. The geometry from Clipper was
considered overlapping and needed to be offsetted by a small margin to
make it non-overalapping.

There is still a bug in the string pulling (most propably): some
short-cuts are taken by the path in some cases.