Added License.txt to 3rdparty/polypartition/
PolyPartition uses MIT license as stated at it's google code page, but
does not have any separate license file distributed with the source files.
The license is found at the start of polypartition.h.

I added the MIT license to sneaky/3rdparty/polypartition and copied the
copyright information from the license in the source file.